OTP Lifecycle

+++ app/src/main/java/crm/sales/com/salescrm/Customer.java
@@ -6,12 +6,12 @@ package crm.sales.com.salescrm;
 
 public class Customer {
 
-    long customerID;
+    int customerID;
     String customerName;
-    public long getCustomerID() {
+    public int getCustomerID() {
         return customerID;
     }
-    public void setCustomerID(long customerID) {
+    public void setCustomerID(int customerID) {
         this.customerID = customerID;
     }
     public String getCustomerName() {
